sharedtopic into seperate file/component
